学习原型模式
